Improve naming to be mor specific
