+++ backend/src/main/java/ch/puzzle/okr/OkrApplication.java
@@ -1,15 +1,14 @@
 package ch.puzzle.okr;
 
-import ch.puzzle.okr.service.clientconfig.ClientConfigProperties;
+import ch.puzzle.okr.service.clientcustomization.ClientCustomizationProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @EnableScheduling
-@EnableConfigurationProperties(ClientConfigProperties.class)
+@EnableConfigurationProperties(ClientCustomizationProperties.class)
 public class OkrApplication {
     public static void main(String[] args) {
         SpringApplication.run(OkrApplication.class, args);
