+++ backend/src/main/resources/application-staging.properties
@@ -4,4 +4,4 @@ logging.level.org.springframework=debug
 spring.security.oauth2.resourceserver.opaquetoken.client-id=pitc_okr_staging
 
 okr.user.champion.usernames=peggimann
-okr.clientconfig.customstyles.okr-topbar-background-color=#ab31ad
+okr.clientcustomization.customstyles.okr-topbar-background-color=#ab31ad
