+++ backend/src/main/resources/application.properties
@@ -51,6 +51,6 @@ okr.jwt.user.firstname=given_name
 okr.jwt.user.lastname=family_name
 okr.jwt.user.email=email
 
-okr.clientconfig.favicon=assets/favicon.png
-okr.clientconfig.logo=assets/images/okr-logo.svg
-okr.clientconfig.title=Puzzle OKR
+okr.clientcustomization.favicon=assets/favicon.png
+okr.clientcustomization.logo=assets/images/okr-logo.svg
+okr.clientcustomization.title=Puzzle OKR
