+++ backend/src/test/java/ch/puzzle/okr/controller/ClientConfigControllerIT.java
@@ -1,6 +1,6 @@
 package ch.puzzle.okr.controller;
 
-import ch.puzzle.okr.service.clientconfig.ClientConfigService;
+import ch.puzzle.okr.service.clientcustomization.ClientConfigService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
